requirements.txt
numpy==1.26.4
scipy==1.10.1
scikit-image==0.21.0
opencv-python==4.7.0.72
Pillow==10.3.0
diffusers==0.27.2
transformers==4.36.2
accelerate==0.26.1
matplotlib==3.7.4
tqdm==4.64.1
gradio==4.16.0
config==0.5.1
einops==0.7.0
onnxruntime==1.16.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
huggingface_hub==0.22.2
insightface==0.7.3
ip_adapter==0.1.0
safetensors
onnxruntime-gpu
omegaconf
open_clip_torch
clip-interrogator